requirements.txt
absl-py                 0.15.0
addict                  2.4.0
aiohttp                 3.8.4
aiosignal               1.2.0
altair                  4.1.0
antlr4-python3-runtime  4.8
anyio                   3.3.4
appdirs                 1.4.4
argon2-cffi             21.3.0
argon2-cffi-bindings    21.2.0
astor                   0.8.1
astunparse              1.6.3
async-generator         1.10
async-timeout           4.0.2
asynctest               0.13.0
attrs                   21.4.0
Babel                   2.9.1
backcall                0.2.0
backports.zoneinfo      0.2.1
base58                  2.1.1
basicsr                 1.3.5
bleach                  4.1.0
blinker                 1.4
cached-property         1.5.2
cachetools              4.2.4
certifi                 2017.4.17
cffi                    1.15.0
chardet                 4.0.0
charset-normalizer      2.0.12
clang                   5.0
clean-fid               0.1.31
click                   8.0.4
clip                    1.0
contextvars             2.4
cycler                  0.11.0
dataclasses             0.8
decorator               4.4.2
deepdiff                5.7.0
defusedxml              0.7.1
deprecation             2.1.0
dill                    0.3.4
dnnlib                  0.0.1
easydict                1.9
einops                  0.4.1
entrypoints             0.4
filelock                3.4.1
flatbuffers             1.12
frozenlist              1.2.0
fsspec                  2022.1.0
ftfy                    6.0.3
future                  0.18.3
fvcore                  0.1.5.post20220305
gast                    0.4.0
gitdb                   4.0.9
GitPython               3.1.20
google-auth             1.35.0
google-auth-oauthlib    0.4.6
google-pasta            0.2.0
grpcio                  1.44.0
h5py                    3.1.0
huggingface-hub         0.4.0
idna                    2.10
idna-ssl                1.1.0
imageio                 2.15.0
imageio-ffmpeg          0.4.5
immutables              0.15
importlib-metadata      4.8.3
importlib-resources     5.4.0
iopath                  0.1.9
ipykernel               5.5.6
ipython                 7.16.3
ipython-genutils        0.2.0
ipywidgets              7.6.5
jedi                    0.17.2
Jinja2                  3.0.3
joblib                  1.1.1
json5                   0.9.5
jsonschema              4.0.0
jupyter-client          7.1.2
jupyter-core            4.9.2
jupyter-packaging       0.10.6
jupyter-server          1.13.1
jupyterlab              3.2.1
jupyterlab-pygments     0.1.2
jupyterlab-server       2.8.2
jupyterlab-widgets      1.0.2
keras                   2.6.0
Keras-Preprocessing     1.1.2
kiwisolver              1.3.1
kornia                  0.6.5
lazy                    1.4
lmdb                    1.3.0
lpips                   0.1.4
Markdown                3.3.6
MarkupSafe              2.0.1
matplotlib              3.3.4
meshio                  4.4.6
mistune                 0.8.4
mrcfile                 1.3.0
multidict               5.2.0
nbclassic               0.3.3
nbclient                0.5.9
nbconvert               6.0.7
nbformat                5.1.3
nest-asyncio            1.5.4
networkx                2.5.1
ninja                   1.10.2.3
notebook                6.4.8
numpy                   1.19.5
oauthlib                3.2.0
omegaconf               2.1.1
open3d                  0.13.0
opencv-python           4.5.5.64
opt-einsum              3.3.0
ordered-set             4.0.2
packaging               21.3
pandas                  1.1.5
pandocfilters           1.5.0
parso                   0.7.1
pexpect                 4.8.0
pickleshare             0.7.5
Pillow                  8.4.0
pip                     21.3.1
plyfile                 0.7.4
portalocker             2.4.0
prometheus-client       0.13.1
prompt-toolkit          3.0.28
protobuf                3.19.4
ptyprocess              0.7.0
pyarrow                 6.0.1
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycparser               2.21
pydeck                  0.6.2
pyDeprecate             0.3.1
Pygments                2.11.2
Pympler                 1.0.1
pyntcloud               0.1.4
pyparsing               3.0.7
pyrsistent              0.18.0
python-dateutil         2.8.2
pytorch-fid             0.2.0
pytorch-lightning       1.5.0
pytz                    2021.3
pytz-deprecation-shim   0.1.0.post0
pyvista                 0.32.1
PyWavelets              1.1.1
PyYAML                  6.0
pyzmq                   22.3.0
regex                   2022.4.24
requests                2.25.1
requests-oauthlib       1.3.1
rsa                     4.8
sacremoses              0.0.53
scikit-image            0.17.2
scikit-learn            0.24.2
scikit-video            1.1.11
scipy                   1.5.4
scooby                  0.5.7
semver                  2.13.0
Send2Trash              1.8.0
setuptools              58.0.4
simdkalman              1.0.2
six                     1.15.0
smmap                   5.0.0
sniffio                 1.2.0
streamlit               1.7.0
tabulate                0.8.9
tb-nightly              2.11.0a20220816
tensorboard             2.6.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.1
tensorflow-estimator    2.6.0
termcolor               1.1.0
terminado               0.13.0
testpath                0.6.0
threadpoolctl           2.1.0
tifffile                2020.9.3
timm                    0.6.7
tl2                     0.0.8
tokenizers              0.12.1
toml                    0.10.2
tomlkit                 0.7.0
toolz                   0.11.2
torch                   1.8.2+cu111
torch-ema               0.2
torch-fidelity          0.2.0
torchmetrics            0.8.2
torchvision             0.9.2+cu111
tornado                 6.1
tqdm                    4.63.0
traitlets               4.3.3
transformers            4.18.0
tsmoothie               1.0.4
typing-extensions       3.7.4.3
tzdata                  2021.5
tzlocal                 4.1
urllib3                 1.26.8
validators              0.18.2
vtk                     9.0.1
watchdog                2.1.6
wcwidth                 0.2.5
webencodings            0.5.1
websocket-client        1.2.3
Werkzeug                2.0.3
wheel                   0.37.1
widgetsnbextension      3.5.2
wrapt                   1.12.1
yacs                    0.1.8
yapf                    0.32.0
yarl                    1.7.2
zipp                    3.6.0
